+++ public/manifest.json
@@ -13,8 +13,9 @@
       "type": "image/png"
     }
   ],
-  "start_url": "./index.html",
+  "start_url": "/",
   "display": "standalone",
-  "background_color": "#000000",
+  "scope": "/",
+  "background_color": "#ffffff",
   "theme_color": "#4DBA87"
 }
