+++ src/assets/styling/scss/layout/_all.scss
@@ -1,4 +1,5 @@
 @import 'grid';
 @import 'layout';
 @import 'header';
+@import 'banner';
 @import 'footer';
