+++ src/assets/styling/scss/pages/_footer.scss
@@ -1,6 +1,7 @@
 .todos-footer {
   @include pxRem("padding", 10 15);
   @include pxRem("height", 50);
+  background-color: white;
   position: relative;
   border-top: 1px solid map-get($colors, border);
   display: flex;
