+++ src/assets/styling/scss/pages/_header.scss
@@ -31,9 +31,10 @@
     align-items: center;
     flex-grow: 12;
     position: fixed;
-    width: 100%;
     top: 0;
     z-index: 10;
+    left: 0;
+    right: 0;
     background: map-get($colors, white);
     border-bottom: 1px solid map-get($colors, border);
     .page-logo {
