+++ src/assets/styling/scss/pages/_home.scss
@@ -9,6 +9,6 @@
   .main {
     @include pxRem("padding-top", 60);
     width: 100%;
-    height: 100vh;
+    min-height: 100vh;
   }
 }
