+++ src/registerServiceWorker.ts
@@ -17,7 +17,7 @@ if (process.env.NODE_ENV === 'production') {
       console.log('Content has been cached for offline use.');
     },
     updatefound() {
-      console.log('New content is downloading.');
+      console.log('New content is downloading.', document.getElementById('banner'));
     },
     updated() {
       console.log('New content is available; please refresh.');
